[#91] refactor into RtpCodec and RtpCodecFactory
