+++ codecs/sip/src/main/java/de/bmarwell/proximo/pitido/codecs/sip/OpusRtpCodec.java
@@ -20,72 +20,18 @@ import java.lang.foreign.MemorySegment;
 import java.lang.foreign.SymbolLookup;
 import java.lang.foreign.ValueLayout;
 import java.lang.invoke.MethodHandle;
-import javax.annotation.PostConstruct;
-import javax.enterprise.context.ApplicationScoped;
 
-/**
- * Opus RTP codec (RFC 7587, dynamic payload type 111).
- *
- * <p>Opus is a versatile open codec developed by the IETF Codec Working Group.
- * At 48 kHz mono it delivers transparent, wideband audio (20 Hz–20 kHz) in VBR mode,
- * making it the codec of choice for FLOSS softphones such as Linphone, Baresip, and Jitsi.
- *
- * <p>Deutsche Telekom, SIPGate, and other PSTN SIP trunks do not support Opus; they will never
- * offer it in an SDP INVITE, so this codec is automatically bypassed for those callers.
- *
- * <h2>SDP declaration</h2>
- *
- * <p>Per RFC 7587 §5, the SDP attribute lines are:
- * <pre>
- * a=rtpmap:111 opus/48000/2
- * a=fmtp:111 useinbandfec=1
- * </pre>
- *
- * <p>The channel count of 2 in the {@code rtpmap} line is mandated by RFC 7587 regardless of
- * the actual channel count in the stream; this codec sends mono.
- *
- * <h2>Encoding parameters</h2>
- *
- * <ul>
- *   <li>Application: {@code OPUS_APPLICATION_VOIP} — optimised for voice; activates
- *       voice activity detection, noise suppression, and in-band FEC.</li>
- *   <li>In-band FEC: enabled via {@code OPUS_SET_INBAND_FEC_REQUEST = 1} — libopus embeds
- *       redundant recovery data for the previous packet inside the current packet, so receivers
- *       can conceal moderate random packet loss without retransmission.</li>
- *   <li>Bitrate: {@code OPUS_AUTO} (default) — libopus selects the optimal VBR bitrate for
- *       the signal content, typically 24–32 kbps for mono voice at 48 kHz.</li>
- *   <li>Frame size: 20 ms (960 samples at 48 kHz) — matches the standard RTP packet cadence.</li>
- * </ul>
- *
- * <h2>Native backend — libopus via FFM</h2>
- *
- * <p>Encoding is performed by {@code libopus} via the Foreign Function and Memory (FFM) API.
- * The same native library is used by
- * {@link de.bmarwell.proximo.pitido.codecs.input.OggOpusPcmDecoder} for audio file decoding;
- * it must be installed on the host system:
- * <ul>
- *   <li>Debian / Ubuntu: {@code apt install libopus0}</li>
- *   <li>Arch Linux: {@code pacman -S opus}</li>
- *   <li>RHEL / UBI 9: {@code microdnf install opus} (available in the base repository)</li>
- * </ul>
- *
- * <h2>Factory / per-call pattern</h2>
- *
- * <p>Opus ADPCM carries predictor state across packets.
- * This {@code @ApplicationScoped} CDI bean acts as a factory: {@link #forCall()} determines the
- * required encoder state size via {@code opus_encoder_get_size}, allocates a confined
- * {@link Arena}, initialises it with {@code opus_encoder_init}, and returns a plain (non-CDI)
- * {@code OpusRtpCodec} instance.
- * When the call ends, {@link de.bmarwell.proximo.pitido.war.media.CallSessionManager} calls
- * {@link #close()}, which releases the confined arena and all native state immediately.
- *
- * @see G722RtpCodec
- * @see NativeRtpCodec
- */
-@ApplicationScoped
-public final class OpusRtpCodec extends NativeRtpCodec {
+public class OpusRtpCodec extends NativeRtpCodec implements RtpCodec {
 
-    private static final System.Logger LOGGER = System.getLogger(OpusRtpCodec.class.getName());
+    /**
+     * Conservative upper bound for one encoded Opus frame.
+     * 4 000 bytes vastly exceeds any realistic 20 ms voice frame;
+     * libopus will never exceed this limit for the given frame size and bitrate.
+     */
+    private static final int MAX_ENCODED_BYTES = 4000;
+
+    /** Mono encoding: one channel. */
+    private static final int OPUS_CHANNELS = 1;
 
     /** libopus application mode optimised for voice; enables DTX, FEC, and noise suppression. */
     private static final int OPUS_APPLICATION_VOIP = 2048;
@@ -96,47 +42,11 @@ public final class OpusRtpCodec extends NativeRtpCodec {
      */
     private static final int OPUS_SET_INBAND_FEC_REQUEST = 4012;
 
-    /** Mono encoding: one channel. */
-    private static final int OPUS_CHANNELS = 1;
-
-    /** Opus native sample rate: 48 000 Hz. */
-    private static final int OPUS_SAMPLE_RATE = 48_000;
-
-    /** 20 ms frame at 48 kHz: 960 samples. */
-    private static final int FRAME_SAMPLES = 960;
-
-    /**
-     * Conservative upper bound for one encoded Opus frame.
-     * 4 000 bytes vastly exceeds any realistic 20 ms voice frame;
-     * libopus will never exceed this limit for the given frame size and bitrate.
-     */
-    private static final int MAX_ENCODED_BYTES = 4000;
-
     /**
      * Alignment for the {@code OpusEncoder} state struct.
      * On x86-64, libopus uses 8-byte-aligned double fields internally.
      */
     private static final long STATE_ALIGN = 8L;
-
-    // -------------------------------------------------------------------------
-    // CDI factory bean — set by @PostConstruct, null in per-call instances
-    // -------------------------------------------------------------------------
-
-    private MethodHandle opusEncoderGetSizeHandle;
-    private MethodHandle opusEncoderInitHandle;
-    /** Typed binding for {@code opus_encoder_ctl(encoder, request, intValue)} (variadic). */
-    private MethodHandle opusEncoderCtlIntHandle;
-
-    // -------------------------------------------------------------------------
-    // Shared between factory and per-call instances
-    // -------------------------------------------------------------------------
-
-    private MethodHandle opusEncodeHandle;
-    private MethodHandle opusEncoderDestroyHandle;
-
-    // -------------------------------------------------------------------------
-    // Per-call instance fields — null in the CDI factory bean
-    // -------------------------------------------------------------------------
 
     /**
      * Pointer to the initialised {@code OpusEncoder} state, allocated inside
@@ -145,169 +55,70 @@ public final class OpusRtpCodec extends NativeRtpCodec {
      */
     private final MemorySegment stateSegment;
 
-    /** CDI no-args constructor. */
-    public OpusRtpCodec() {
-        this.stateSegment = null;
-    }
+    private final int sampleRate;
+    private final int frameSamples;
 
-    /**
-     * Per-call constructor — creates a non-CDI encoder instance for exactly one call leg.
-     *
-     * <p>Not intended for direct use; called only by {@link #forCall()}.
-     *
-     * @param opusEncodeHandle        downcall handle for {@code opus_encode}
-     * @param opusEncoderDestroyHandle downcall handle for {@code opus_encoder_destroy}
-     * @param callArena               confined arena owning the encoder state
-     * @param stateSegment            initialised {@code OpusEncoder} state
-     */
-    OpusRtpCodec(
-            MethodHandle opusEncodeHandle,
-            MethodHandle opusEncoderDestroyHandle,
-            Arena callArena,
-            MemorySegment stateSegment) {
-        super(callArena);
-        this.opusEncodeHandle = opusEncodeHandle;
-        this.opusEncoderDestroyHandle = opusEncoderDestroyHandle;
-        this.stateSegment = stateSegment;
-    }
+    private MethodHandle opusEncodeHandle;
+    private MethodHandle opusEncoderDestroyHandle;
 
-    /**
-     * Probes for {@code libopus.so.0} and binds all required FFM method handles.
-     *
-     * <p>Called once by the CDI container after construction.
-     * Sets {@link NativeRtpCodec#available} to {@code true} when the library is found.
-     * Uses {@link Arena#global()} so the library remains loaded for the lifetime of the JVM.
-     */
-    @PostConstruct
-    @SuppressWarnings("restricted") // SymbolLookup.libraryLookup is FFM restricted — intentional use
-    void probe() {
-        try {
-            SymbolLookup opus = SymbolLookup.libraryLookup("libopus.so.0", Arena.global());
-            Linker linker = Linker.nativeLinker();
+    private MethodHandle opusEncoderGetSizeHandle;
+    private MethodHandle opusEncoderInitHandle;
+    /** Typed binding for {@code opus_encoder_ctl(encoder, request, intValue)} (variadic). */
+    private MethodHandle opusEncoderCtlIntHandle;
 
-            this.opusEncoderGetSizeHandle = linker.downcallHandle(
-                    opus.find("opus_encoder_get_size").orElseThrow(),
-                    FunctionDescriptor.of(
-                            ValueLayout.JAVA_INT, // return: size in bytes
-                            ValueLayout.JAVA_INT // channels
-                            ));
+    protected OpusRtpCodec(int opusSampleRate, int frameSamples) {
+        this.sampleRate = opusSampleRate;
+        this.frameSamples = frameSamples;
+        SymbolLookup opus = SymbolLookup.libraryLookup("libopus.so.0", Arena.global());
+        Linker linker = Linker.nativeLinker();
 
-            this.opusEncoderInitHandle = linker.downcallHandle(
-                    opus.find("opus_encoder_init").orElseThrow(),
-                    FunctionDescriptor.of(
-                            ValueLayout.JAVA_INT, // return: OPUS_OK (0) on success
-                            ValueLayout.ADDRESS, // OpusEncoder* st
-                            ValueLayout.JAVA_INT, // Fs (sample rate)
-                            ValueLayout.JAVA_INT, // channels
-                            ValueLayout.JAVA_INT // application
-                            ));
+        this.opusEncoderGetSizeHandle = linker.downcallHandle(
+                opus.find("opus_encoder_get_size").orElseThrow(),
+                FunctionDescriptor.of(
+                        ValueLayout.JAVA_INT, // return: size in bytes
+                        ValueLayout.JAVA_INT // channels
+                        ));
 
-            this.opusEncodeHandle = linker.downcallHandle(
-                    opus.find("opus_encode").orElseThrow(),
-                    FunctionDescriptor.of(
-                            ValueLayout.JAVA_INT, // return: bytes encoded (>0) or error (<0)
-                            ValueLayout.ADDRESS, // OpusEncoder* st
-                            ValueLayout.ADDRESS, // const opus_int16* pcm
-                            ValueLayout.JAVA_INT, // frame_size (samples)
-                            ValueLayout.ADDRESS, // unsigned char* data (output)
-                            ValueLayout.JAVA_INT // max_data_bytes
-                            ));
+        this.opusEncoderInitHandle = linker.downcallHandle(
+                opus.find("opus_encoder_init").orElseThrow(),
+                FunctionDescriptor.of(
+                        ValueLayout.JAVA_INT, // return: OPUS_OK (0) on success
+                        ValueLayout.ADDRESS, // OpusEncoder* st
+                        ValueLayout.JAVA_INT, // Fs (sample rate)
+                        ValueLayout.JAVA_INT, // channels
+                        ValueLayout.JAVA_INT // application
+                        ));
 
-            this.opusEncoderDestroyHandle = linker.downcallHandle(
-                    opus.find("opus_encoder_destroy").orElseThrow(), FunctionDescriptor.ofVoid(ValueLayout.ADDRESS));
+        this.opusEncodeHandle = linker.downcallHandle(
+                opus.find("opus_encode").orElseThrow(),
+                FunctionDescriptor.of(
+                        ValueLayout.JAVA_INT, // return: bytes encoded (>0) or error (<0)
+                        ValueLayout.ADDRESS, // OpusEncoder* st
+                        ValueLayout.ADDRESS, // const opus_int16* pcm
+                        ValueLayout.JAVA_INT, // frame_size (samples)
+                        ValueLayout.ADDRESS, // unsigned char* data (output)
+                        ValueLayout.JAVA_INT // max_data_bytes
+                        ));
 
-            // Variadic binding for opus_encoder_ctl(encoder, request, int_value).
-            // firstVariadicArg(2) tells the linker that argument index 2 is the first variadic.
-            this.opusEncoderCtlIntHandle = linker.downcallHandle(
-                    opus.find("opus_encoder_ctl").orElseThrow(),
-                    FunctionDescriptor.of(
-                            ValueLayout.JAVA_INT, // return: OPUS_OK (0)
-                            ValueLayout.ADDRESS, // OpusEncoder* st
-                            ValueLayout.JAVA_INT, // request
-                            ValueLayout.JAVA_INT // variadic int value
-                            ),
-                    Linker.Option.firstVariadicArg(2));
+        this.opusEncoderDestroyHandle = linker.downcallHandle(
+                opus.find("opus_encoder_destroy").orElseThrow(), FunctionDescriptor.ofVoid(ValueLayout.ADDRESS));
 
-            this.available = true;
-            LOGGER.log(System.Logger.Level.INFO, "libopus detected — Opus RTP codec available");
+        // Variadic binding for opus_encoder_ctl(encoder, request, int_value).
+        // firstVariadicArg(2) tells the linker that argument index 2 is the first variadic.
+        this.opusEncoderCtlIntHandle = linker.downcallHandle(
+                opus.find("opus_encoder_ctl").orElseThrow(),
+                FunctionDescriptor.of(
+                        ValueLayout.JAVA_INT, // return: OPUS_OK (0)
+                        ValueLayout.ADDRESS, // OpusEncoder* st
+                        ValueLayout.JAVA_INT, // request
+                        ValueLayout.JAVA_INT // variadic int value
+                        ),
+                Linker.Option.firstVariadicArg(2));
 
-        } catch (IllegalArgumentException illegalArgumentException) {
-            LOGGER.log(
-                    System.Logger.Level.WARNING,
-                    "libopus not found — Opus RTP codec disabled: {0}",
-                    illegalArgumentException.getMessage());
-        }
-    }
-
-    @Override
-    public int preference() {
-        // Preferred over G.722 (50) when both are available: Opus delivers higher quality
-        // and wider frequency response at equivalent or lower bitrate.
-        // PSTN trunks never offer Opus, so this preference only activates for softphone callers.
-        return 30;
-    }
-
-    /**
-     * Returns a new per-call encoder instance with freshly initialised Opus encoder state.
-     *
-     * <p>Uses {@code opus_encoder_get_size(1)} to determine the required allocation, then
-     * allocates a confined arena and calls {@code opus_encoder_init} followed by
-     * {@code opus_encoder_ctl} to enable in-band FEC.
-     *
-     * @throws IllegalStateException if encoder initialisation fails
-     */
-    @Override
-    public RtpCodecFactory forCall() {
         int stateSize = invokeGetSize();
-        Arena arena = Arena.ofConfined();
-        MemorySegment state = arena.allocate(stateSize, STATE_ALIGN);
-        invokeInit(state);
-        invokeCtlInt(state, OPUS_SET_INBAND_FEC_REQUEST, 1);
-
-        return new OpusRtpCodec(this.opusEncodeHandle, this.opusEncoderDestroyHandle, arena, state);
-    }
-
-    @Override
-    public int payloadType() {
-        // De-facto standard dynamic payload type for Opus, used by virtually all softphones.
-        return 111;
-    }
-
-    @Override
-    public int rtpClockRate() {
-        // RFC 7587 §4: Opus RTP clock rate is 48 000 Hz (no historical anomaly here).
-        return OPUS_SAMPLE_RATE;
-    }
-
-    @Override
-    public int inputSampleRate() {
-        return OPUS_SAMPLE_RATE;
-    }
-
-    @Override
-    public int samplesPerFrame() {
-        // 20 ms × 48 000 Hz = 960 samples
-        return FRAME_SAMPLES;
-    }
-
-    @Override
-    public int rtpTimestampIncrement() {
-        // 48 000 Hz / 50 packets per second = 960
-        return FRAME_SAMPLES;
-    }
-
-    @Override
-    public String sdpName() {
-        return "opus";
-    }
-
-    /**
-     * Returns {@code 2} as required by RFC 7587 §5, regardless of actual channel count.
-     * The stream is mono; the SDP channel field is fixed at 2 for interoperability.
-     */
-    @Override
-    public int sdpChannelCount() {
-        return 2;
+        this.stateSegment = callArena.allocate(stateSize, STATE_ALIGN);
+        invokeInit(this.stateSegment);
+        invokeCtlInt(this.stateSegment, OPUS_SET_INBAND_FEC_REQUEST, 1);
     }
 
     @Override
@@ -321,10 +132,8 @@ public final class OpusRtpCodec extends NativeRtpCodec {
      *
      * <p>The returned byte array length varies per frame (VBR); the caller wraps it directly
      * as the RTP payload.
-     * This method is only valid on per-call instances created by {@link #forCall()}.
      *
-     * @param pcmFrame 960 mono PCM samples at 48 000 Hz; length must equal
-     *                 {@link #samplesPerFrame()}
+     * @param pcmFrame 960 mono PCM samples at 48 000 Hz; length must equal {@link #frameSamples}
      * @return Opus-encoded bytes for one RTP packet
      * @throws IOException           if {@code opus_encode} returns a negative error code
      * @throws IllegalStateException if called on the CDI factory bean (no encoder state)
@@ -350,6 +159,17 @@ public final class OpusRtpCodec extends NativeRtpCodec {
         }
     }
 
+    private int invokeEncode(MemorySegment inputSeg, MemorySegment outputSeg) throws IOException {
+        try {
+            return (int) this.opusEncodeHandle.invoke(
+                    this.stateSegment, inputSeg, this.frameSamples, outputSeg, MAX_ENCODED_BYTES);
+        } catch (RuntimeException runtimeException) {
+            throw runtimeException;
+        } catch (Throwable throwable) {
+            throw new IOException("opus_encode invocation failed", throwable);
+        }
+    }
+
     private int invokeGetSize() {
         try {
             return (int) this.opusEncoderGetSizeHandle.invoke(OPUS_CHANNELS);
@@ -363,7 +183,7 @@ public final class OpusRtpCodec extends NativeRtpCodec {
     private void invokeInit(MemorySegment state) {
         try {
             int result = (int)
-                    this.opusEncoderInitHandle.invoke(state, OPUS_SAMPLE_RATE, OPUS_CHANNELS, OPUS_APPLICATION_VOIP);
+                    this.opusEncoderInitHandle.invoke(state, this.sampleRate, OPUS_CHANNELS, OPUS_APPLICATION_VOIP);
 
             if (result != 0) {
                 throw new IllegalStateException("opus_encoder_init failed with error code " + result);
@@ -382,17 +202,6 @@ public final class OpusRtpCodec extends NativeRtpCodec {
             throw runtimeException;
         } catch (Throwable throwable) {
             throw new IllegalStateException("opus_encoder_ctl invocation failed", throwable);
-        }
-    }
-
-    private int invokeEncode(MemorySegment inputSeg, MemorySegment outputSeg) throws IOException {
-        try {
-            return (int) this.opusEncodeHandle.invoke(
-                    this.stateSegment, inputSeg, FRAME_SAMPLES, outputSeg, MAX_ENCODED_BYTES);
-        } catch (RuntimeException runtimeException) {
-            throw runtimeException;
-        } catch (Throwable throwable) {
-            throw new IOException("opus_encode invocation failed", throwable);
         }
     }
 }
