+++ codecs/sip/src/test/java/de/bmarwell/proximo/pitido/codecs/sip/OpusRtpCodecTest.java
@@ -13,7 +13,6 @@
 package de.bmarwell.proximo.pitido.codecs.sip;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
@@ -21,76 +20,11 @@ import org.junit.jupiter.api.Test;
 
 class OpusRtpCodecTest {
 
-    private final OpusRtpCodec codec = new OpusRtpCodec();
-
-    // -------------------------------------------------------------------------
-    // Codec constants — no native library needed
-    // -------------------------------------------------------------------------
-
-    @Test
-    void payloadTypeIs111() {
-        assertEquals(111, codec.payloadType());
-    }
-
-    @Test
-    void rtpClockRateIs48000() {
-        assertEquals(48_000, codec.rtpClockRate());
-    }
-
-    @Test
-    void inputSampleRateIs48000() {
-        assertEquals(48_000, codec.inputSampleRate());
-    }
-
-    @Test
-    void samplesPerFrameIs960() {
-        assertEquals(960, codec.samplesPerFrame());
-    }
-
-    @Test
-    void rtpTimestampIncrementIs960() {
-        assertEquals(960, codec.rtpTimestampIncrement());
-    }
-
-    @Test
-    void sdpNameIsOpus() {
-        assertEquals("opus", codec.sdpName());
-    }
-
-    @Test
-    void sdpChannelCountIsTwo() {
-        // given: RFC 7587 §5 mandates 2 in SDP regardless of actual channel count
-        // when / then
-        assertEquals(2, codec.sdpChannelCount());
-    }
+    private final OpusRtpCodec codec = (OpusRtpCodec) new OpusRtpCodecFactory().forCall("");
 
     @Test
     void fmtpParamsContainsFec() {
         assertEquals("useinbandfec=1", codec.fmtpParams());
-    }
-
-    @Test
-    void preferenceIs30() {
-        assertEquals(30, codec.preference());
-    }
-
-    @Test
-    void preferenceIsHigherThanG722() {
-        // given
-        var g722 = new G722RtpCodec();
-
-        // when / then: Opus (30) must beat G.722 (50) — lower number = higher priority
-        assertEquals(
-                true,
-                codec.preference() < g722.preference(),
-                "Opus preference must be higher priority (lower number) than G.722");
-    }
-
-    @Test
-    void encodeThrowsOnFactoryBean() {
-        // given: an un-probed CDI factory bean with no encoder state
-        // when / then
-        assertThrows(IllegalStateException.class, () -> codec.encode(new short[960]));
     }
 
     // -------------------------------------------------------------------------
@@ -99,53 +33,22 @@ class OpusRtpCodecTest {
 
     @Test
     void closeOnFactoryBean_isNoOp() {
-        // given: a fresh factory bean (no callArena)
-        var factory = new OpusRtpCodec();
+        // given: a fresh factory
+        var factory = new OpusRtpCodecFactory().forCall("");
 
-        // when / then: close() must not throw on the CDI factory bean
+        // when / then: close() must not throw
         factory.close();
-    }
-
-    @Test
-    void forCallReturnsDifferentInstance() {
-        // given
-        OpusRtpCodec factory = new OpusRtpCodec();
-        factory.probe();
-
-        assumeTrue(factory.isAvailable(), "libopus not available on this host — skipping");
-
-        // when
-        RtpCodecFactory callInstance = factory.forCall();
-
-        // then
-        assertNotSame(factory, callInstance);
-    }
-
-    @Test
-    void forCallReturnsFreshInstanceEachTime() {
-        // given
-        OpusRtpCodec factory = new OpusRtpCodec();
-        factory.probe();
-
-        assumeTrue(factory.isAvailable(), "libopus not available on this host — skipping");
-
-        // when
-        RtpCodecFactory first = factory.forCall();
-        RtpCodecFactory second = factory.forCall();
-
-        // then: each call leg must have independent encoder state
-        assertNotSame(first, second);
     }
 
     @Test
     void closeOnPerCallInstance_releasesArena() {
         // given
-        OpusRtpCodec factory = new OpusRtpCodec();
+        OpusRtpCodecFactory factory = new OpusRtpCodecFactory();
         factory.probe();
 
         assumeTrue(factory.isAvailable(), "libopus not available on this host — skipping");
 
-        RtpCodecFactory callInstance = factory.forCall();
+        RtpCodec callInstance = factory.forCall("");
 
         // when
         callInstance.close();
@@ -157,12 +60,12 @@ class OpusRtpCodecTest {
     @Test
     void silenceFrameEncodesSuccessfully() throws Exception {
         // given
-        OpusRtpCodec factory = new OpusRtpCodec();
+        OpusRtpCodecFactory factory = new OpusRtpCodecFactory();
         factory.probe();
 
         assumeTrue(factory.isAvailable(), "libopus not available on this host — skipping");
 
-        RtpCodecFactory encoder = factory.forCall();
+        RtpCodec encoder = factory.forCall("");
         short[] silence = new short[960];
 
         // when
